Update Dockerfile
Updated the docker file to include the latest version of the debian-based DW Spectrum installer package. This is necessary because those running Windows 10 or 11 clients of DW Spectrum face massive stability and crashing issues when attempting to connect to a host running the 4.x.x. client from 2021. This will create the container with the latest version as of Feb 18th. 2025 (6.0.0.39503)